+++ src/harlequin/cli.py
@@ -1,5 +1,4 @@
-from pathlib import Path
-from typing import List
+from typing import List, Union
 
 import click
 
@@ -26,13 +25,37 @@ from harlequin import Harlequin
         "monokai."
     ),
 )
+@click.option(
+    "--md_token",
+    help=(
+        "MotherDuck Token. Pass your MotherDuck service token in this option, or "
+        "set the `motherduck_token` environment variable."
+    ),
+)
+@click.option(
+    "--md_saas",
+    is_flag=True,
+    help="Run MotherDuck in SaaS mode (no local privileges).",
+)
 @click.argument(
     "db_path",
     nargs=-1,
-    type=click.Path(path_type=Path),
+    type=click.Path(path_type=str),
 )
-def harlequin(db_path: List[Path], read_only: bool, theme: str) -> None:
+def harlequin(
+    db_path: List[str],
+    read_only: bool,
+    theme: str,
+    md_token: Union[str, None],
+    md_saas: bool,
+) -> None:
     if not db_path:
-        db_path = [Path(":memory:")]
-    tui = Harlequin(db_path=db_path, read_only=read_only, theme=theme)
+        db_path = [":memory:"]
+    tui = Harlequin(
+        db_path=db_path,
+        read_only=read_only,
+        theme=theme,
+        md_token=md_token,
+        md_saas=md_saas,
+    )
     tui.run()